mouse manager : vector2f in world

+++ MouseManager.h
@@ -5,13 +5,13 @@ class MouseManager
 {
 
 private:
-    sf::Vector2i clickPosition;
+    sf::Vector2f clickPosition;
     bool isClicked = false;
 
 public:
     void update(sf::Event event, sf::RenderWindow& window);
 
-    sf::Vector2i getClickPosition();
+    sf::Vector2f getClickPosition();
     bool isMouseClicked();
 
 };
